fix normalization of paths in Galley service (#57)
I've noticed that normally a path for REST API requests usually
does not starts with / (because the path spec is `/v1/{path=**}`.

It could come to a confusion if paths in gRPC requests start with
a slash. Thus Galley server normalizes the path to ensure that
the path does not start with slashes at least for the internal
form.

On the other hand, etcd storage adopter has another wrong
normalization -- / is added to the key conditionally, which
also complicates the watching or getting the responses.

So on the etcd side, it should always add the slash when a key
comes in and and remove the slash when the key goes out.